Новые стили

+++ app/Http/Controllers/EmployeeController.php
@@ -4,11 +4,16 @@ namespace App\Http\Controllers;
 
 use App\Models\Employee;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class EmployeeController extends Controller
 {
     public function index(Request $request)
     {
+        if (!Auth::user()->hasRole('Role_ADMIN')) {
+            return redirect()->route('home')->with('error', 'У вас нет доступа к этой странице.');
+        }
+
         $query = Employee::query();
 
         // Фильтрация
@@ -34,12 +39,20 @@ class EmployeeController extends Controller
 
     public function create(Request $request)
     {
+        if (!Auth::user()->hasRole('Role_ADMIN')) {
+            return redirect()->route('home')->with('error', 'У вас нет доступа к этой странице.');
+        }
+
         $request->session()->forget('current_section');
         return view('employees.create');
     }
 
     public function store(Request $request)
     {
+        if (!Auth::user()->hasRole('Role_ADMIN')) {
+            return redirect()->route('home')->with('error', 'У вас нет доступа к этой странице.');
+        }
+
         $validatedData = $request->validate([
             'lastName' => 'required|string|max:255',
             'firstName' => 'required|string|max:255',
@@ -56,12 +69,20 @@ class EmployeeController extends Controller
 
     public function edit(Request $request, Employee $employee)
     {
+        if (!Auth::user()->hasRole('Role_ADMIN')) {
+            return redirect()->route('home')->with('error', 'У вас нет доступа к этой странице.');
+        }
+
         $request->session()->forget('current_section');
         return view('employees.edit', compact('employee'));
     }
 
     public function update(Request $request, Employee $employee)
     {
+        if (!Auth::user()->hasRole('Role_ADMIN')) {
+            return redirect()->route('home')->with('error', 'У вас нет доступа к этой странице.');
+        }
+
         $validatedData = $request->validate([
             'lastName' => 'required|string|max:255',
             'firstName' => 'required|string|max:255',
@@ -78,6 +99,10 @@ class EmployeeController extends Controller
 
     public function destroy(Request $request, Employee $employee)
     {
+        if (!Auth::user()->hasRole('Role_ADMIN')) {
+            return redirect()->route('home')->with('error', 'У вас нет доступа к этой странице.');
+        }
+
         $employee->delete();
 
         $request->session()->forget('current_section');
